Re-write lesson subject code
I still can't verify whether this code is fully correct because I
haven't been able to find any sources of historical data about which
subjects fell on which dates. In particular, I'm not sure whether the
extra lesson subject (in a year with 53 Sundays) is always added at the
end of the year or whether it can sometimes occur mid-year.